requirements.txt
groq
selenium
webdriver-manager
googlesearch-python
cohere
requests
bs4
pygame==2.5.2
edge-tts==7.0.0
PyQt5
appopener
python-dotenv
pillow
pywhatkit
rich
mtranslate
keyboard
pyaudio==0.2.11
numpy==1.24.3
psutil==5.9.5
